Rename BlacklightRangeLimit::FacetFieldComponent to RangeFieldComponent
